sched: compute task utilisation with WALT consistently
Using WALT, the utilisation of a task is computed with a resolution
scaling factor that has been used inconsistently in the code with either
hardcoded values or macros (NICE_0_LOAD_SHIFT in this case). Changes in
these macros (as the 32 to 64 bits resolution shift of 2159197d66770)
happened to break the utilisation calculation wherever they have been
used whilst results remained correct in other places. This commit fixes
this issue by using SCHED_CAPACITY_SCALE as resolution scaling factor
consistently.

Change-Id: Ic5418f8a5dfc455a22bafbebb4142b4665b61c6f
Signed-off-by: Quentin Perret <quentin.perret@arm.com>

Conflicts:
	kernel/sched/fair.c
Signed-off-by: FAROVITUS <farovitus@gmail.com>