Revert "dts: Fix a bug were during a reboot, system could freeze at splash screen"
Fix broken MTP connection, this mean that we have the reboot bug again.

If anyone have an idea about it or wants to experiment it more, the bug is related to 'vdd_ldo12', so adding back 'regulator-always-on;' will also fix the reboot bug but that's not a healthy solution.
Let me know if any of found something new about this situation.

This reverts commit de18b9a8f1b1fe66f808c8d037b40ef0fc9ca9c8.